fix: generate png-url while convert full color elements
